+++ css/src/icon-extractor-tools.ts
@@ -31,14 +31,14 @@ COLOR_PREFIXES.forEach((prefix) => {
             // always keep the group-focus and group-hover classes
             return [`icon-${normalClass}`, `group-${normalClass}`]
           }
-    }
+    },
   )
 })
 
 export { ICON_ATTRIBUTE_NAMES_TO_CLASS_GENERATOR }
 
 export function isIconAttribute(
-  attrName: string
+  attrName: string,
 ): attrName is keyof typeof ICON_ATTRIBUTE_NAMES_TO_CLASS_GENERATOR {
   return ICON_ATTRIBUTE_NAMES_TO_CLASS_GENERATOR[attrName] !== undefined
 }