+++ css/src/tw-icon-extractor.test.ts
@@ -31,7 +31,7 @@ describe('getHtmlAttributes', () => {
 describe('IconExtractor', () => {
   it('should extract colors', () => {
     const ext = IconExtractor(
-      '<wind-keep fillColor="red-300" strokeColor="indigo-800" />'
+      '<wind-keep fillColor="red-300" strokeColor="indigo-800" />',
     )
     expect(ext).toContain('icon-light-red-300')
     expect(ext).toContain('icon-dark-indigo-800')
@@ -39,7 +39,7 @@ describe('IconExtractor', () => {
 
   it('should extract colors in curlies', () => {
     const ext = IconExtractor(
-      "<wind-keep fillColor={'red-300'} strokeColor={'indigo-800'} />"
+      "<wind-keep fillColor={'red-300'} strokeColor={'indigo-800'} />",
     )
     expect(ext).toContain('icon-light-red-300')
     expect(ext).toContain('icon-dark-indigo-800')