+++ components/DocMenu/vue/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @cypress-design/vue-docmenu
 
+## 0.6.1
+
+### Patch Changes
+
+- [`ad32042`](https://github.com/cypress-io/cypress-design/commit/ad3204289730de696c9728b1343e06ce68382561) Thanks [@elevatebart](https://github.com/elevatebart)! - only skip scrolling if both dimension are in view
+
+## 0.6.0
+
+### Minor Changes
+
+- [#388](https://github.com/cypress-io/cypress-design/pull/388) [`7839c67`](https://github.com/cypress-io/cypress-design/commit/7839c670d8a83f8339db1a5d4395bd0dd5eaf987) Thanks [@elevatebart](https://github.com/elevatebart)! - make docmenu scroll to active path if necessary
+
+## 0.5.3
+
+### Patch Changes
+
+- [#387](https://github.com/cypress-io/cypress-design/pull/387) [`05b67c5`](https://github.com/cypress-io/cypress-design/commit/05b67c5389df1c0619d2f01e2b36fb5bd6bb30b9) Thanks [@elevatebart](https://github.com/elevatebart)! - when activePath changes, expand sections that contain the newly activated element
+
 ## 0.5.2
 
 ### Patch Changes